backend/requirements.txt
# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.12.0

# Caching / queuing
redis>=5.0.0
celery>=5.3.0
aioredis>=2.0.0

# Bioinformatics
biopython>=1.81

# Chemoinformatics (install via: pip install rdkit-pypi OR conda install -c conda-forge rdkit)
rdkit-pypi>=2022.9.5

# ML / Data science
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
shap>=0.43.0

# HTTP client
httpx>=0.25.0
aiohttp>=3.9.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Retry / resilience
tenacity>=8.2.0

# Logging
structlog>=23.2.0

# Metrics
prometheus-client>=0.18.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0
